pyproject.toml
[project]
name = "md-tts"
version = "0.1.0"
description = "CLI to convert Markdown into TTS audio (English/Spanish)"
authors = [{ name = "Equipo DMBOK", email = "example@example.com" }]
requires-python = ">=3.10, <=3.12"
dependencies = [
    "transformers>=4.46.0",
    "accelerate>=0.30.0",
    "huggingface_hub>=0.24.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "numpy>=1.25.0",
    "tqdm>=4.66.1",
    "kokoro>=0.9.2",
    "chatterbox-tts>=0.1.6",
]

[project.scripts]
md-tts = "md_tts.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
